+++ CHANGELOG.md
@@ -21,6 +21,14 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+## v0.1.0
+
+Released Thursday, 20th February 2020.
+
+### Backwards-Compatibility Breaks
+
+* Updated error callbacks to use `@ganbarodigital/ts-lib-error-reporting`
+
 ## v0.0.2
 
 Released Wednesday, 19th February 2020.